requirements.txt
flask
gunicorn
psycopg2-binary
requests
python-dotenv
pandas
numpy
pytz
